+++ recipes/mosquitto/2.x/CMakeLists.txt
@@ -4,4 +4,8 @@ project(conan_cmake_wrapper)
 include(conanbuildinfo.cmake)
 conan_basic_setup()
 
+if (WIN32 AND BUILD_SHARED_LIBS)
+    set(CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS ON)
+endif()
+
 add_subdirectory(source_subfolder)
